+++ girder_annotation/girder_large_image_annotation/web_client/views/itemList.js
@@ -1,11 +1,11 @@
 import $ from 'jquery';
 import _ from 'underscore';
 
+import { restRequest } from '@girder/core/rest';
 import { wrap } from '@girder/core/utilities/PluginUtils';
 import ItemListWidget from '@girder/core/views/widgets/ItemListWidget';
 
 import largeImageAnnotationConfig from './configView';
-import AnnotationCollection from '../collections/AnnotationCollection';
 
 import '../stylesheets/itemList.styl';
 
@@ -13,10 +13,16 @@ wrap(ItemListWidget, 'render', function (render) {
     render.apply(this, _.rest(arguments));
 
     function addLargeImageAnnotationBadge(item, parent) {
-        const collection = new AnnotationCollection([]);
-        collection.fetch({
-            itemId: item.id
-        }).done(() => {
+        restRequest({
+            type: 'GET',
+            url: 'annotation',
+            data: {
+                itemId: item.id,
+                limit: -1
+            },
+            error: null
+        }).done((result, status, jqxhr) => {
+            const numAnnotations = +jqxhr.getResponseHeader('Girder-Total-Count') || 0;
             const thumbnail = $('a[g-item-cid="' + item.cid + '"] .large_image_thumbnail', parent).first();
 
             let badge = thumbnail.find('.large_image_annotation_badge');
@@ -24,9 +30,8 @@ wrap(ItemListWidget, 'render', function (render) {
                 badge = $(`<div class="large_image_annotation_badge"></div>`).appendTo(thumbnail);
             }
             // update badge
-            const numAnnotations = collection.length;
             badge
-                .attr('title', `${numAnnotations} Annotation${numAnnotations === 1 ? '' : 's'}`)
+                .attr('title', `${numAnnotations} annotation${numAnnotations === 1 ? '' : 's'}`)
                 .text(numAnnotations)
                 .toggleClass('hidden', numAnnotations === 0);
         });